doc: fix doc warnings
